+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
-## 0.1.0 (2023-12-31)
+## 0.1.3 (2023-03-07)
+- roster UI with shift creation and team member assignment considering their availability (without data storage)
+- absence UI with absence creation 
+- contact data UI
+- work agreements UI
+- user settings UI (without data storage)
+- document menu with ChangeHistory
+
+## 0.1.0 (2023-10-27)
 
 - app setup for multiple platforms with electron and cordova
 - color scheme and app icons